install/kubernetes: add helm option to override lbmap max entries
This commits adds a new helm bpf option `global.bpf.lbMapMax` that can
be used to configure the max entries in LB service, backend and affinity
maps. Extend KubeProxy free gsg to point out the use of this helm
option when user needs to scale the cluster beyond a default preconfigured
limit.

Signed-off-by: Deepesh Pathak <deepshpathak@gmail.com>